[Fix] code change
